+++ src/app/shared/components/shop-creation/create-russian-shop-entity/create-russian-shop-entity.module.ts
@@ -23,7 +23,7 @@ import { DaDataModule } from '../../../../dadata';
 import { ShopPayoutToolDetailsService } from '../../../../sections/payment-section/integrations/shops/services/shop-payout-tool-details/shop-payout-tool-details.service';
 import { ExistingBankAccountComponent } from './components/existing-bank-account/existing-bank-account.component';
 import { OrgDetailsFormComponent } from './components/org-details-form/org-details-form.component';
-import { RussianBankAccountFormComponent } from './components/russian-bank-account/russian-bank-account-form.component';
+import { RussianBankAccountFormComponent } from './components/russian-bank-account-form/russian-bank-account-form.component';
 import { ShopFormComponent } from './components/shop-form/shop-form.component';
 import { CreateRussianShopEntityComponent } from './create-russian-shop-entity.component';
 import { CreateRussianShopEntityService } from './services/create-russian-shop-entity/create-russian-shop-entity.service';
